standalone_bria/requirements_batch.txt
# Requirements for batch_background_remover.py
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
scikit-image>=0.19.0